% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{get_raw_directory}
\alias{get_raw_directory}
\title{Get raw school directory data from ADE Report Cards API}
\usage{
get_raw_directory(end_year, include_contact = TRUE)
}
\arguments{
\item{end_year}{Fiscal year end}

\item{include_contact}{Whether to fetch contact details}
}
\value{
Raw data frame from API
}
\description{
Downloads the raw school directory data from the Arizona Department of
Education Report Cards API.
}
\keyword{internal}
