% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{merge_safely}
\alias{merge_safely}
\title{Safely merge two dataframes on entity ID}
\usage{
merge_safely(base, new_data, entity_col)
}
\arguments{
\item{base}{Base dataframe}

\item{new_data}{New dataframe to merge}

\item{entity_col}{Name of entity ID column}
}
\value{
Merged dataframe
}
\description{
Safely merge two dataframes on entity ID
}
\keyword{internal}
