% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{parse_era1_excel}
\alias{parse_era1_excel}
\title{Parse Era 1 Excel file (2011-2017)}
\usage{
parse_era1_excel(excel_path, end_year)
}
\arguments{
\item{excel_path}{Path to downloaded Excel file}

\item{end_year}{School year end}
}
\value{
List with school and district data frames
}
\description{
Older ADE enrollment files have multiple tabs for different levels.
}
\keyword{internal}
