% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{read_and_merge_sheets}
\alias{read_and_merge_sheets}
\title{Read and merge enrollment sheets}
\usage{
read_and_merge_sheets(excel_path, sheets, level)
}
\arguments{
\item{excel_path}{Path to Excel file}

\item{sheets}{Vector of sheet names}

\item{level}{Either "school" or "lea"}
}
\value{
Merged data frame
}
\description{
Reads data from Grade, Gender, Ethnicity, and Subgroup sheets
and merges them into a single wide-format dataframe.
}
\keyword{internal}
